fix: PR issues addressed
